services/resizing/requirements.txt
aiohttp==3.3.2
Pillow==5.2.0